refactor: fold tracing tests into their unit test files (CL2-962)
The tracing/span assertions lived in dedicated *_tracing_test.rb files,
split from each unit's behavior tests. That split was accepted while the
tracing epic was in flight; the preference is one test file per unit.

Move each standalone file's contents into its unit test file under a
nested `describe "tracing"` block and delete the standalone file:

  run_tracing_test.rb            -> run_test.rb
  providers/chat_tracing_test.rb -> providers/base_test.rb
  tools/runtime_tracing_test.rb  -> tools/runtime_test.rb

Each `describe "tracing"` scopes its own before/after so the in-memory
tracer install and the tracing.enabled / capture_messages toggles never
leak onto the file's non-tracing tests, and keeps the
`skip unless OTEL_SDK_AVAILABLE` guard so the no-OTEL CI lane stays green.

Cosmetic sub-describes are flattened into the tracing block; groups whose
before hook is load-bearing (chat content-capture, runtime span shape /
returned error responses / raised exceptions) stay nested. In runtime,
the duplicate weather_tool_class is dropped in favour of the parent's and
make_tool_call is unified with an `arguments: "{}"` default. Two chat
client-span tests are renamed (generate_text / stream_text) to avoid a
duplicate test name once flattened.
